[v2.1.3] OSS Merge

+++ app/packages/spotlight/src/row.ts
@@ -48,8 +48,8 @@ export default class Row<K, V> {
       element.style.top = pixels(ZERO);
 
       if (config.onItemClick) {
-        const handler = (event) => {
-          if (event.metaKey || event.shiftKey) {
+        const handler = (event: MouseEvent) => {
+          if (event.metaKey || event.shiftKey || event.ctrlKey) {
             return;
           }
 
@@ -119,8 +119,8 @@ export default class Row<K, V> {
     return this.#row[this.#row.length - ONE].item.id;
   }
 
-  destroy() {
-    this.#destroyItems();
+  destroy(destroyItems = false) {
+    destroyItems && this.#destroyItems();
     this.#aborter.abort();
   }
 
@@ -138,8 +138,11 @@ export default class Row<K, V> {
       throw new Error("row is not attached");
     }
 
+    if (!this.#config.retainItems) {
+      this.#destroyItems();
+    }
+
     this.#container.remove();
-    this.#destroyItems();
   }
 
   show(
